fronend headlamp-plugin: Use source in lib/ rather than types/
This also adds all of the headlamp dependencies
into headlamp-plugin.

As before, `npm run build` will compile the source.
This time it is compiled into the lib/ folder, rather than types/
Additionally, this time the .js source files are in there along
with the type declarations in lib/.

The reason to compile to .js and include type declarations is so
that people can still use JavaScript if they want. Additionally
the configuration and setup seems to be better supported this way.

types/ folder is removed and all that reference it cease to exist.

@remix-run/router had to be pinned at 1.0.1 because of typescript
errors in 1.0.3 (issue reported). I expect these will go away with
a react-router v6 update.

The .prettierignore is there to ignore the generated .js files.